+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 2.13a (XX.XX.XXXX)
 ------------------
 
-* TBA
+* Dropped support for ``relative_url()`` methods on custom menu item models that do not support a ``request`` keyword argument.
+
 
 
 2.12 (17.11.2018)